[front] Extract model render preparation
